More PDP Logging and Minor Lift Changes
PDP Logs the current state (Auton, Teleop, Disabled), and will forcibly
run once on the disabed init so you can see when it is disabled.

The lift climb setpoints have been adjusted, and now it won't switch to
starting flip if the new setpoint is higher than the starting height.